feat: Participants 삭제 및 메서드 일부 이동
